Add sample code.
